+++ services/horizon/internal/actions/account_test.go
@@ -4,17 +4,115 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+
 	protocol "github.com/stellar/go/protocols/horizon"
 	"github.com/stellar/go/services/horizon/internal/db2/core"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/test"
+	"github.com/stellar/go/support/render/problem"
+	"github.com/stellar/go/xdr"
+)
+
+var (
+	trustLineIssuer = "GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H"
+	accountOne      = "GABGMPEKKDWR2WFH5AJOZV5PDKLJEHGCR3Q24ALETWR5H3A7GI3YTS7V"
+	accountTwo      = "GADTXHUTHIAESMMQ2ZWSTIIGBZRLHUCBLCHPLLUEIAWDEFRDC4SYDKOZ"
+	accountThree    = "GDP347UYM2ZKE6ED6T5OM3BQ5IAS76NKRVEUPNB5PCQ26Z5D7Q7PJOMI"
+	signer          = "GCXKG6RN4ONIEPCMNFB732A436Z5PNDSRLGWK7GBLCMQLIFO4S7EYWVU"
+	usd             = xdr.MustNewCreditAsset("USD", trustLineIssuer)
+	euro            = xdr.MustNewCreditAsset("EUR", trustLineIssuer)
+
+	account1 = xdr.AccountEntry{
+		AccountId:     xdr.MustAddress(accountOne),
+		Balance:       20000,
+		SeqNum:        223456789,
+		NumSubEntries: 10,
+		Flags:         1,
+		HomeDomain:    "stellar.org",
+		Thresholds:    xdr.Thresholds{1, 2, 3, 4},
+		Ext: xdr.AccountEntryExt{
+			V: 1,
+			V1: &xdr.AccountEntryV1{
+				Liabilities: xdr.Liabilities{
+					Buying:  3,
+					Selling: 4,
+				},
+			},
+		},
+	}
+
+	account2 = xdr.AccountEntry{
+		AccountId:     xdr.MustAddress(accountTwo),
+		Balance:       50000,
+		SeqNum:        648736,
+		NumSubEntries: 10,
+		Flags:         2,
+		HomeDomain:    "meridian.stellar.org",
+		Thresholds:    xdr.Thresholds{5, 6, 7, 8},
+		Ext: xdr.AccountEntryExt{
+			V: 1,
+			V1: &xdr.AccountEntryV1{
+				Liabilities: xdr.Liabilities{
+					Buying:  30,
+					Selling: 40,
+				},
+			},
+		},
+	}
+	eurTrustLine = xdr.TrustLineEntry{
+		AccountId: xdr.MustAddress(accountOne),
+		Asset:     euro,
+		Balance:   20000,
+		Limit:     223456789,
+		Flags:     1,
+		Ext: xdr.TrustLineEntryExt{
+			V: 1,
+			V1: &xdr.TrustLineEntryV1{
+				Liabilities: xdr.Liabilities{
+					Buying:  3,
+					Selling: 4,
+				},
+			},
+		},
+	}
+
+	usdTrustLine = xdr.TrustLineEntry{
+		AccountId: xdr.MustAddress(accountTwo),
+		Asset:     usd,
+		Balance:   10000,
+		Limit:     123456789,
+		Flags:     0,
+		Ext: xdr.TrustLineEntryExt{
+			V: 1,
+			V1: &xdr.TrustLineEntryV1{
+				Liabilities: xdr.Liabilities{
+					Buying:  1,
+					Selling: 2,
+				},
+			},
+		},
+	}
+
+	data1 = xdr.DataEntry{
+		AccountId: xdr.MustAddress(accountOne),
+		DataName:  "test data",
+		// This also tests if base64 encoding is working as 0 is invalid UTF-8 byte
+		DataValue: []byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9},
+	}
+
+	data2 = xdr.DataEntry{
+		AccountId: xdr.MustAddress(accountTwo),
+		DataName:  "test data2",
+		DataValue: []byte{10, 11, 12, 13, 14, 15, 16, 17, 18, 19},
+	}
 )
 
 func TestAccountInfo(t *testing.T) {
 	tt := test.Start(t).Scenario("allow_trust")
 	defer tt.Finish()
 
-	account, err := AccountInfo(tt.Ctx, &core.Q{tt.CoreSession()}, "GCXKG6RN4ONIEPCMNFB732A436Z5PNDSRLGWK7GBLCMQLIFO4S7EYWVU")
+	account, err := AccountInfo(tt.Ctx, &core.Q{tt.CoreSession()}, signer)
 	tt.Assert.NoError(err)
 
 	tt.Assert.Equal("8589934593", account.Sequence)
@@ -34,13 +132,13 @@ func TestGetAccountsHandlerPageNoResults(t *testing.T) {
 	test.ResetHorizonDB(t, tt.HorizonDB)
 
 	q := &history.Q{tt.HorizonSession()}
-	handler := &GetAccountsHandler{HistoryQ: q}
+	handler := &GetAccountsHandler{}
 	records, err := handler.GetResourcePage(
 		httptest.NewRecorder(),
 		makeRequest(
 			t,
 			map[string]string{
-				"signer": "GCXKG6RN4ONIEPCMNFB732A436Z5PNDSRLGWK7GBLCMQLIFO4S7EYWVU",
+				"signer": signer,
 			},
 			map[string]string{},
 			q.Session,
@@ -50,31 +148,15 @@ func TestGetAccountsHandlerPageNoResults(t *testing.T) {
 	tt.Assert.Len(records, 0)
 }
 
-func TestGetAccountsHandlerPageResults(t *testing.T) {
+func TestGetAccountsHandlerPageResultsBySigner(t *testing.T) {
 	tt := test.Start(t)
 	defer tt.Finish()
 	test.ResetHorizonDB(t, tt.HorizonDB)
 
 	q := &history.Q{tt.HorizonSession()}
-	handler := &GetAccountsHandler{HistoryQ: q}
+	handler := &GetAccountsHandler{}
 
-	rows := []history.AccountSigner{
-		history.AccountSigner{
-			Account: "GABGMPEKKDWR2WFH5AJOZV5PDKLJEHGCR3Q24ALETWR5H3A7GI3YTS7V",
-			Signer:  "GCXKG6RN4ONIEPCMNFB732A436Z5PNDSRLGWK7GBLCMQLIFO4S7EYWVU",
-			Weight:  1,
-		},
-		history.AccountSigner{
-			Account: "GADTXHUTHIAESMMQ2ZWSTIIGBZRLHUCBLCHPLLUEIAWDEFRDC4SYDKOZ",
-			Signer:  "GCXKG6RN4ONIEPCMNFB732A436Z5PNDSRLGWK7GBLCMQLIFO4S7EYWVU",
-			Weight:  2,
-		},
-		history.AccountSigner{
-			Account: "GDP347UYM2ZKE6ED6T5OM3BQ5IAS76NKRVEUPNB5PCQ26Z5D7Q7PJOMI",
-			Signer:  "GCXKG6RN4ONIEPCMNFB732A436Z5PNDSRLGWK7GBLCMQLIFO4S7EYWVU",
-			Weight:  3,
-		},
-	}
+	rows := accountSigners()
 
 	for _, row := range rows {
 		q.CreateAccountSigner(row.Account, row.Signer, row.Weight)
@@ -85,7 +167,7 @@ func TestGetAccountsHandlerPageResults(t *testing.T) {
 		makeRequest(
 			t,
 			map[string]string{
-				"signer": "GCXKG6RN4ONIEPCMNFB732A436Z5PNDSRLGWK7GBLCMQLIFO4S7EYWVU",
+				"signer": signer,
 			},
 			map[string]string{},
 			q.Session,
@@ -123,5 +205,173 @@ func TestGetAccountsHandlerPageResults(t *testing.T) {
 		tt.Assert.Equal(row.Account, result.AccountID)
 		tt.Assert.Equal(row.Signer, result.Signer.Key)
 		tt.Assert.Equal(row.Weight, result.Signer.Weight)
+	}
+}
+
+func TestGetAccountsHandlerPageResultsByAsset(t *testing.T) {
+	tt := test.Start(t)
+	defer tt.Finish()
+	test.ResetHorizonDB(t, tt.HorizonDB)
+
+	q := &history.Q{tt.HorizonSession()}
+	handler := &GetAccountsHandler{}
+
+	_, err := q.InsertAccount(account1, 1234)
+	tt.Assert.NoError(err)
+	_, err = q.InsertAccount(account2, 1234)
+	tt.Assert.NoError(err)
+
+	rows := accountSigners()
+
+	for _, row := range rows {
+		_, err = q.CreateAccountSigner(row.Account, row.Signer, row.Weight)
+		tt.Assert.NoError(err)
+	}
+
+	_, err = q.InsertAccountData(data1, 1234)
+	assert.NoError(t, err)
+	_, err = q.InsertAccountData(data2, 1234)
+	assert.NoError(t, err)
+
+	var assetType, code, issuer string
+	usd.MustExtract(&assetType, &code, &issuer)
+	params := map[string]string{
+		"asset_issuer": issuer,
+		"asset_code":   code,
+		"asset_type":   assetType,
+	}
+
+	records, err := handler.GetResourcePage(
+		httptest.NewRecorder(),
+		makeRequest(
+			t,
+			params,
+			map[string]string{},
+			q.Session,
+		),
+	)
+
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(0, len(records))
+
+	_, err = q.InsertTrustLine(eurTrustLine, 1234)
+	assert.NoError(t, err)
+	_, err = q.InsertTrustLine(usdTrustLine, 1235)
+	assert.NoError(t, err)
+
+	records, err = handler.GetResourcePage(
+		httptest.NewRecorder(),
+		makeRequest(
+			t,
+			params,
+			map[string]string{},
+			q.Session,
+		),
+	)
+
+	tt.Assert.NoError(err)
+	tt.Assert.Equal(1, len(records))
+	result := records[0].(protocol.Account)
+	tt.Assert.Equal(accountTwo, result.AccountID)
+	tt.Assert.Len(result.Balances, 2)
+	tt.Assert.Len(result.Signers, 2)
+
+	_, ok := result.Data[string(data2.DataName)]
+	tt.Assert.True(ok)
+}
+
+func TestGetAccountsHandlerInvalidParams(t *testing.T) {
+	testCases := []struct {
+		desc                    string
+		params                  map[string]string
+		expectedInvalidField    string
+		expectedErr             string
+		isInvalidAccountsParams bool
+	}{
+		{
+			desc:                    "empty filters",
+			isInvalidAccountsParams: true,
+		},
+		{
+			desc: "signer and seller",
+			params: map[string]string{
+				"signer":       accountOne,
+				"asset_issuer": accountOne,
+				"asset_code":   "USD",
+				"asset_type":   "credit_alphanum4",
+			},
+			expectedInvalidField: "signer",
+			expectedErr:          "you can't filter by signer and asset at the same time",
+		},
+		{
+			desc: "filtering by native asset",
+			params: map[string]string{
+				"asset_type": "native",
+			},
+			expectedInvalidField: "asset_type",
+			expectedErr:          "you can't filter by asset type: native",
+		},
+		{
+			desc: "invalid asset",
+			params: map[string]string{
+				"asset_issuer": accountOne,
+				"asset_code":   "USDCOP",
+				"asset_type":   "credit_alphanum4",
+			},
+			expectedInvalidField: "asset_code",
+			expectedErr:          "Asset code must be 1-12 alphanumeric characters",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.desc, func(t *testing.T) {
+			tt := test.Start(t)
+			defer tt.Finish()
+			q := &history.Q{tt.HorizonSession()}
+			handler := &GetAccountsHandler{}
+
+			_, err := handler.GetResourcePage(
+				httptest.NewRecorder(),
+				makeRequest(
+					t,
+					tc.params,
+					map[string]string{},
+					q.Session,
+				),
+			)
+			tt.Assert.Error(err)
+			if tc.isInvalidAccountsParams {
+				tt.Assert.Equal(invalidAccountsParams, err)
+			} else {
+				if tt.Assert.IsType(&problem.P{}, err) {
+					p := err.(*problem.P)
+					tt.Assert.Equal("bad_request", p.Type)
+					tt.Assert.Equal(tc.expectedInvalidField, p.Extras["invalid_field"])
+					tt.Assert.Equal(
+						tc.expectedErr,
+						p.Extras["reason"],
+					)
+				}
+			}
+		})
+	}
+}
+
+func accountSigners() []history.AccountSigner {
+	return []history.AccountSigner{
+		history.AccountSigner{
+			Account: accountOne,
+			Signer:  signer,
+			Weight:  1,
+		},
+		history.AccountSigner{
+			Account: accountTwo,
+			Signer:  signer,
+			Weight:  2,
+		},
+		history.AccountSigner{
+			Account: accountThree,
+			Signer:  signer,
+			Weight:  3,
+		},
 	}
 }